tetris: extract background.h
